+++ tests/mocks/mockApiServer.mjs
@@ -78,6 +78,35 @@ app.get("/:key", (req, res) => {
             ErrorCode: 105,
             ErrorMessage: "Other error occurred",
         });
+    // Handle submission data retrieval
+    } else if (key === "submissionData") {
+        return res.status(200).json({
+            Succeeded: true,
+            ErrorCode: 0,
+            ErrorMessage: null,
+            Data: {
+                submissionId: "f97c091b-caa6-4e69-a515-f57043a704b0",
+                serviceId: "f11077f3-72c7-4eec-ab01-f07d9efed08d",
+                referenceValue: "0000000924107836",
+                submissionTs: "2025-08-07T17:05:26+03:00",
+                status: 0,
+                statusTs: "2025-08-07T17:05:34+03:00",
+                userId: "0000000924107836",
+                userType: 1,
+                userName: "",
+                userIdentifier: "0001013317",
+                userEmail: "",
+                submissionData: "{\"index\":{\"formData\":{\"certificate_select\":[\"birth\",\"permanent_residence\"]}},\"data-entry-radios\":{\"formData\":{\"mobile_select\":\"other\",\"mobileTxt\":\"+35799484967\"}}}",
+                dataStatus: 0
+            }
+        });
+    } else if (key === "submissionEmpty") {
+        return res.status(200).json({
+            Succeeded: true,
+            ErrorCode: 0,
+            ErrorMessage: null,
+            Data: null
+        });
     } else {
         return res.status(400).json({
             Succeeded: false,
@@ -87,7 +116,90 @@ app.get("/:key", (req, res) => {
     }
 });
 
-// Start the server
+
+// Mock endpoint to download a file
+app.get("/:key/:fileId/:sha256", (req, res) => {
+    const { key, fileId, sha256 } = req.params;
+    const authHeader = req.headers['authorization'] || null;
+    console.log(`Received request with key: ${key}`);
+    console.log(`Authorization header: ${authHeader}`);
+    console.log(`Query params: ${JSON.stringify(req.query, null, 2)}`); 
+    console.log(`Request body: ${JSON.stringify(req.body, null, 2)}`);
+
+    // Simulate different responses based on input
+    if (key === "fileDownload") {
+        return res.status(200).json({
+            Succeeded: true,
+            ErrorCode: 0,
+            ErrorMessage: null,
+            Data : {
+                fileId: "685ce92aa0291b778956dea3",
+                fileName: "govcy (21).pdf",
+                contentType: "application/pdf",
+                fileSize: 1872,
+                sha256: "bb89074e5374c96a6c2a9ea2e6b7f5885693edf6d06c3dadbdb41b5d7a44253c",
+                base64: "JVBERi0xLjMKJZOMi54gUmVwb3J0TGFiIEdlbmVyYXRlZCBQREYgZG9jdW1lbnQgaHR0cDovL3d3dy5yZXBvcnRsYWIuY29tCjEgMCBvYmoKPDwKL0YxIDIgMCBSCj4+CmVuZG9iagoyIDAgb2JqCjw8Ci9CYXNlRm9udCAvSGVsdmV0aWNhIC9FbmNvZGluZyAvV2luQW5zaUVuY29kaW5nIC9OYW1lIC9GMSAvU3VidHlwZSAvVHlwZTEgL1R5cGUgL0ZvbnQKPj4KZW5kb2JqCjMgMCBvYmoKPDwKL0NvbnRlbnRzIDcgMCBSIC9NZWRpYUJveCBbIDAgMCA2MTIgNzkyIF0gL1BhcmVudCA2IDAgUiAvUmVzb3VyY2VzIDw8Ci9Gb250IDEgMCBSIC9Qcm9jU2V0IFsgL1BERiAvVGV4dCAvSW1hZ2VCIC9JbWFnZUMgL0ltYWdlSSBdCj4+IC9Sb3RhdGUgMCAvVHJhbnMgPDwKCj4+IAogIC9UeXBlIC9QYWdlCj4+CmVuZG9iago0IDAgb2JqCjw8Ci9QYWdlTW9kZSAvVXNlTm9uZSAvUGFnZXMgNiAwIFIgL1R5cGUgL0NhdGFsb2cKPj4KZW5kb2JqCjUgMCBvYmoKPDwKL0F1dGhvciAoYW5vbnltb3VzKSAvQ3JlYXRpb25EYXRlIChEOjIwMjUwODA3MDUxNzAyKzAwJzAwJykgL0NyZWF0b3IgKFJlcG9ydExhYiBQREYgTGlicmFyeSAtIHd3dy5yZXBvcnRsYWIuY29tKSAvS2V5d29yZHMgKCkgL01vZERhdGUgKEQ6MjAyNTA4MDcwNTE3MDIrMDAnMDAnKSAvUHJvZHVjZXIgKFJlcG9ydExhYiBQREYgTGlicmFyeSAtIHd3dy5yZXBvcnRsYWIuY29tKSAKICAvU3ViamVjdCAodW5zcGVjaWZpZWQpIC9UaXRsZSAodW50aXRsZWQpIC9UcmFwcGVkIC9GYWxzZQo+PgplbmRvYmoKNiAwIG9iago8PAovQ291bnQgMSAvS2lkcyBbIDMgMCBSIF0gL1R5cGUgL1BhZ2VzCj4+CmVuZG9iago3IDAgb2JqCjw8Ci9GaWx0ZXIgWyAvQVNDSUk4NURlY29kZSAvRmxhdGVEZWNvZGUgXSAvTGVuZ3RoIDEwMgo+PgpzdHJlYW0KR2FwUWgwRT1GLDBVXEgzVFxwTllUXlFLaz90Yz5JUCw7VyNVMV4yM2loUEVNXz9DVzRLSVNpOTBNakdeMixGUyM8UkM1K2MsbigvI11PZTQ8IV5URCNnaV0mPFQhb1lsM1MoK34+ZW5kc3RyZWFtCmVuZG9iagp4cmVmCjAgOAowMDAwMDAwMDAwIDY1NTM1IGYgCjAwMDAwMDAwNzMgMDAwMDAgbiAKMDAwMDAwMDEwNCAwMDAwMCBuIAowMDAwMDAwMjExIDAwMDAwIG4gCjAwMDAwMDA0MDQgMDAwMDAgbiAKMDAwMDAwMDQ3MiAwMDAwMCBuIAowMDAwMDAwNzY4IDAwMDAwIG4gCjAwMDAwMDA4MjcgMDAwMDAgbiAKdHJhaWxlcgo8PAovSUQgCls8ZGMwMzIwYWRmODcxZTliYmY2ODk4NzIxMWUyYmQ3ODI+PGRjMDMyMGFkZjg3MWU5YmJmNjg5ODcyMTFlMmJkNzgyPl0KJSBSZXBvcnRMYWIgZ2VuZXJhdGVkIFBERiBkb2N1bWVudCAtLSBkaWdlc3QgKGh0dHA6Ly93d3cucmVwb3J0bGFiLmNvbSkKCi9JbmZvIDUgMCBSCi9Sb290IDQgMCBSCi9TaXplIDgKPj4Kc3RhcnR4cmVmCjEwMTkKJSVFT0YK",
+                description: null,
+                uid: null,
+                tag: "kostis"
+            },
+            ReceivedFileId: fileId, // This is to test the fileId
+            ReceivedSha256: sha256 // This is to test the sha256
+        });
+    } else if (key === "fileDownloadError401") {
+        return res.status(401).json({
+            Succeeded: false,
+            ErrorCode: 401,
+            ErrorMessage: "Unauthorized access",
+        });
+    } else if (key === "fileDownloadError404") {
+        return res.status(404).json({
+            Succeeded: false,
+            ErrorCode: 404,
+            ErrorMessage: "File not found",
+        });
+    }else {
+        return res.status(400).json({
+            Succeeded: false,
+            ErrorCode: 400,
+            ErrorMessage: "Bad request",
+        });
+    }
+});
+
+// Mock put endpoint for submission
+app.put("/:key", (req, res) => {
+    const { key } = req.params;
+    const authHeader = req.headers['authorization'] || null;
+    console.log(`Received request with key: ${key}`);
+    console.log(`Authorization header: ${authHeader}`);
+    console.log(`Request body: ${JSON.stringify(req.body, null, 2)}`);
+
+    // Simulate different responses based on input
+    if (key === "submissionData") {
+        return res.status(200).json({
+            succeeded: true,
+            errorCode: 0,
+            errorMessage: null,
+            data : { 
+                referenceValue: "0000000924107836" 
+            },
+          });
+    } else if (key === "error401") {
+        return res.status(401).json({
+            Succeeded: false,
+            ErrorCode: 401,
+            ErrorMessage: "Unauthorized access",
+        });
+    }  else {
+        return res.status(400).json({
+            Succeeded: false,
+            ErrorCode: 400,
+            ErrorMessage: "Bad request",
+        });
+    }
+});
+
 app.listen(port, () => {
     console.log(`Mock API running at http://localhost:${port}`);
 });